app/requirement.txt
fastapi==0.95.0
tortoise-orm==0.20.0
uvicorn==0.22.0
python-dotenv==0.20.0
numpy==1.24.1
face_recognition==1.3.0
dotenv